@@ -10,7 +10,8 @@
     "@babel/polyfill": "^7.6.0",
     "@expo/react-native-action-sheet": "^3.7.0",
     "@ptomasroos/react-native-multi-slider": "^1.0.0",
-    "@react-native-community/datetimepicker": "^2.4.0",
+    "@react-native-community/datetimepicker": "3.0.4",
+    "@react-native-community/slider": "^3.0.3",
     "add": "^2.0.6",
     "axios": "^0.18.1",
     "color": "^3.1.2",
@@ -18,14 +19,16 @@
     "cubic-spline": "^2.0.0",
     "data-forge": "^1.7.3",
     "everpolate": "0.0.3",
-    "expo": "^36.0.0",
-    "expo-application": "~2.0.0",
-    "expo-camera": "^8.2.0",
-    "expo-device": "~2.1.0",
-    "expo-image-manipulator": "^8.1.0",
-    "expo-image-picker": "^8.1.0",
-    "expo-network": "~2.1.0",
+    "expo": "^39.0.5",
+    "expo-application": "~2.4.1",
+    "expo-camera": "~9.1.0",
+    "expo-device": "~2.4.0",
+    "expo-image-manipulator": "~8.4.0",
+    "expo-image-picker": "~9.2.0",
+    "expo-network": "~2.4.0",
+    "expo-notifications": "~0.7.2",
     "expo-victory": "^32.0.1",
+    "expo-web-browser": "~8.6.0",
     "firebase": "7.9.0",
     "form-data": "^2.5.1",
     "fs": "0.0.1-security",
@@ -38,34 +41,31 @@
     "numeral": "^2.0.6",
     "qs": "^6.8.0",
     "querystring": "^0.2.0",
-    "react": "16.9.0",
-    "react-dom": "16.9.0",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-37.0.1.tar.gz",
+    "react": "16.13.1",
+    "react-dom": "16.13.1",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-40.0.0.tar.gz",
     "react-native-action-sheet": "^2.2.0",
     "react-native-alert-async": "git+https://github.com/slorber/react-native-alert-async.git",
     "react-native-collapse-view": "^1.0.0",
     "react-native-collapsible": "^1.5.1",
     "react-native-foldview": "^1.2.1",
     "react-native-fs": "^2.16.6",
-    "react-native-gesture-handler": "~1.6.0",
+    "react-native-gesture-handler": "~1.8.0",
     "react-native-gifted-chat": "^0.13.0",
     "react-native-htmlview": "^0.16.0",
     "react-native-keyboard-aware-scroll-view": "^0.8.0",
     "react-native-masked-text": "^1.13.0",
-    "react-native-material-dropdown": "^0.11.1",
-    "react-native-modal-dropdown": "^0.7.0",
     "react-native-pdf": "^6.0.1",
     "react-native-progress": "^3.6.0",
     "react-native-progress-wheel": "^1.0.4",
-    "react-native-reanimated": "~1.7.0",
+    "react-native-reanimated": "~1.13.0",
     "react-native-render-html": "^4.2.0",
-    "react-native-screens": "~2.2.0",
+    "react-native-screens": "~2.15.0",
     "react-native-simple-radio-button": "^2.7.3",
-    "react-native-slider": "^0.11.0",
     "react-native-snap-carousel": "^3.8.1",
     "react-native-step-indicator": "0.0.11",
-    "react-native-svg": "9.13.3",
-    "react-native-svg-transformer": "^0.14.2",
+    "react-native-svg": "12.1.0",
+    "react-native-svg-transformer": "^0.14.3",
     "react-native-swipe-list-view": "^2.0.1",
     "react-native-switch-selector": "^2.0.0",
     "react-native-tab-view": "2.11.0",
@@ -74,8 +74,8 @@
     "react-native-vector-icons": "^6.6.0",
     "react-native-vertical-tab-view": "^0.1.3",
     "react-native-view-pdf": "^0.10.1",
-    "react-native-web": "^0.11.7",
-    "react-native-webview": "8.1.1",
+    "react-native-web": "~0.13.12",
+    "react-native-webview": "11.0.0",
     "react-navigation": "^3.12.1",
     "react-navigation-hooks": "^1.1.0",
     "react-navigation-tabs": "^2.7.0",
@@ -86,7 +86,7 @@
     "yarn": "^1.17.3"
   },
   "devDependencies": {
-    "babel-preset-expo": "^8.0.0"
+    "babel-preset-expo": "8.3.0"
   },
   "private": true
 }
