@@ -21,7 +21,8 @@ class FadeInLeft extends React.Component {
         toValue: 0,
         duration : 1000,
         easing: Easing.elastic(),
-        speed : 1
+        speed : 1,
+        useNativeDriver: true,
       }
     ).start()
   }
