@@ -278,7 +278,8 @@ class AdminScreen extends React.Component {
                                       toValue: 0,
                                       duration : 1000,
                                       easing: Easing.elastic(),
-                                      speed : 1
+                                      speed : 1,
+                                      useNativeDriver: true,
                                     }
                               ),
                                 /*Animated.timing(
@@ -327,7 +328,8 @@ class AdminScreen extends React.Component {
                                                   toValue: getConstant('width'),
                                                   duration : 1000,
                                                   easing: Easing.elastic(),
-                                                  speed : 1
+                                                  speed : 1,
+                                                  useNativeDriver: true,
                                                 }
                                           ),
                                             /*Animated.timing(
